+++ packages/stream_chat_flutter/lib/src/message_action/message_actions_builder.dart
@@ -35,19 +35,19 @@ class StreamMessageActionsBuilder {
         value: ResendMessage(message: message),
         label: Text(context.translations.sendAnywayLabel),
         leading: Icon(
-          icons.send20,
+          icons.send,
           color: StreamChatTheme.of(context).colorTheme.accentPrimary,
         ),
       ),
       StreamContextMenuAction<MessageAction>(
         value: EditMessage(message: message),
         label: Text(context.translations.editMessageLabel),
-        leading: Icon(icons.edit20),
+        leading: Icon(icons.edit),
       ),
       StreamContextMenuAction<MessageAction>.destructive(
         value: HardDeleteMessage(message: message),
         label: Text(context.translations.deleteMessageLabel),
-        leading: Icon(icons.delete20),
+        leading: Icon(icons.delete),
       ),
     ];
   }
@@ -77,7 +77,7 @@ class StreamMessageActionsBuilder {
         if (messageState.isSendingFailed || messageState.isUpdatingFailed) ...[
           StreamContextMenuAction(
             value: ResendMessage(message: message),
-            leading: Icon(icons.send20),
+            leading: Icon(icons.send),
             label: Text(
               context.translations.toggleResendOrResendEditedMessage(
                 isUpdateFailed: messageState.isUpdatingFailed,
@@ -87,7 +87,7 @@ class StreamMessageActionsBuilder {
           if (messageState.isSendingFailed)
             StreamContextMenuAction.destructive(
               value: HardDeleteMessage(message: message),
-              leading: Icon(icons.delete20),
+              leading: Icon(icons.delete),
               label: Text(
                 context.translations.toggleDeleteRetryDeleteMessageText(
                   isDeleteFailed: false,
@@ -98,7 +98,7 @@ class StreamMessageActionsBuilder {
         if (message.state.isDeletingFailed)
           StreamContextMenuAction.destructive(
             value: ResendMessage(message: message),
-            leading: Icon(icons.delete20),
+            leading: Icon(icons.delete),
             label: Text(
               context.translations.toggleDeleteRetryDeleteMessageText(
                 isDeleteFailed: true,
@@ -133,7 +133,7 @@ class StreamMessageActionsBuilder {
         StreamContextMenuAction(
           value: QuotedReply(message: message),
           label: Text(context.translations.replyLabel),
-          leading: Icon(icons.reply20),
+          leading: Icon(icons.reply),
         ),
       );
     }
@@ -143,7 +143,7 @@ class StreamMessageActionsBuilder {
         StreamContextMenuAction(
           value: ThreadReply(message: message),
           label: Text(context.translations.threadReplyLabel),
-          leading: Icon(icons.thread20),
+          leading: Icon(icons.thread),
         ),
       );
     }
@@ -154,7 +154,7 @@ class StreamMessageActionsBuilder {
         return StreamContextMenuAction(
           value: MarkUnread(message: message),
           label: Text(context.translations.markAsUnreadLabel),
-          leading: Icon(icons.notification20),
+          leading: Icon(icons.notification),
         );
       }
 
@@ -175,7 +175,7 @@ class StreamMessageActionsBuilder {
         StreamContextMenuAction(
           value: CopyMessage(message: message),
           label: Text(context.translations.copyMessageLabel),
-          leading: Icon(icons.copy20),
+          leading: Icon(icons.copy),
         ),
       );
     }
@@ -186,7 +186,7 @@ class StreamMessageActionsBuilder {
           StreamContextMenuAction(
             value: EditMessage(message: message),
             label: Text(context.translations.editMessageLabel),
-            leading: Icon(icons.edit20),
+            leading: Icon(icons.edit),
           ),
         );
       }
@@ -208,7 +208,7 @@ class StreamMessageActionsBuilder {
         StreamContextMenuAction(
           value: action,
           label: Text(label.call(pinned: isPinned)),
-          leading: Icon(icons.pin20),
+          leading: Icon(icons.pin),
         ),
       );
     }
@@ -219,7 +219,7 @@ class StreamMessageActionsBuilder {
       messageActions.add(
         StreamContextMenuAction.destructive(
           value: DeleteMessage(message: message),
-          leading: Icon(icons.delete20),
+          leading: Icon(icons.delete),
           label: Text(label.call(isDeleteFailed: false)),
         ),
       );
@@ -230,7 +230,7 @@ class StreamMessageActionsBuilder {
         StreamContextMenuAction(
           value: FlagMessage(message: message),
           label: Text(context.translations.flagMessageLabel),
-          leading: Icon(icons.flag20),
+          leading: Icon(icons.flag),
         ),
       );
     }
@@ -249,7 +249,7 @@ class StreamMessageActionsBuilder {
         StreamContextMenuAction(
           value: action,
           label: Text(label.call(isMuted: isMuted)),
-          leading: Icon(icons.mute20),
+          leading: Icon(icons.mute),
         ),
       );
     }