+++ sample_app/lib/pages/channel_file_display_screen.dart
@@ -1,7 +1,6 @@
 // ignore_for_file: deprecated_member_use
 
 import 'package:flutter/material.dart';
-import 'package:sample_app/utils/localizations.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
 import 'package:video_player/video_player.dart';
 
@@ -43,7 +42,7 @@ class _ChannelFileDisplayScreenState extends State<ChannelFileDisplayScreen> {
         elevation: 1,
         centerTitle: true,
         title: Text(
-          AppLocalizations.of(context).files,
+          'Files',
           style: TextStyle(
             color: StreamChatTheme.of(context).colorTheme.textHighEmphasis,
             fontSize: 16,
@@ -68,13 +67,13 @@ class _ChannelFileDisplayScreenState extends State<ChannelFileDisplayScreen> {
                         mainAxisAlignment: MainAxisAlignment.center,
                         children: [
                           Icon(
-                            context.streamIcons.file32,
+                            context.streamIcons.file,
                             size: 136,
                             color: StreamChatTheme.of(context).colorTheme.disabled,
                           ),
                           const SizedBox(height: 16),
                           Text(
-                            AppLocalizations.of(context).noFiles,
+                            'No Files',
                             style: TextStyle(
                               fontSize: 14,
                               color: StreamChatTheme.of(context).colorTheme.textHighEmphasis,
@@ -82,7 +81,7 @@ class _ChannelFileDisplayScreenState extends State<ChannelFileDisplayScreen> {
                           ),
                           const SizedBox(height: 8),
                           Text(
-                            AppLocalizations.of(context).filesAppearHere,
+                            'Files sent in this chat will appear here',
                             textAlign: TextAlign.center,
                             style: TextStyle(
                               fontSize: 14,